requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Hugging Face Hub (for model downloads)
huggingface-hub>=0.19.0

# Optional: For fine-tuning (uncomment if needed)
# peft>=0.7.0
# bitsandbytes>=0.41.0
# datasets>=2.14.0
